Avalon

+++ Avalon/Avalon.Text/WriteArg.cs
@@ -19,7 +19,7 @@ public class WriteArg : Any
     public virtual long Base { get; set; }
     public virtual long Case { get; set; }
     public virtual long Sign { get; set; }
-    public virtual long FillChar { get; set; }
+    public virtual uint FillChar { get; set; }
     public virtual long ValueCount { get; set; }
     public virtual long Count { get; set; }
 }
